- decomp2json - ajuste em importVAVOEZ para contornar problema ao exportar para JSON - documentação do processo de importação do DECOMP - ajuste do setup para o pip reconhecer comando decomp2json

+++ deckparser/importers/decomp/importVAZOES.py
@@ -8,30 +8,30 @@ def importVAZOES(fn,blockSize):
     vazaoData = fromfile(fn, dtype=vazaoDtype, sep="")
 
     # registro 1 - número de postos, número de estágios e aberturas de cada estágio
-    vazoes["numPostos"] = vazaoData[0]['dado']
-    vazoes["numEstagios"] = vazaoData[1]['dado']
+    vazoes["numPostos"] = int(vazaoData[0]['dado'])
+    vazoes["numEstagios"] = int(vazaoData[1]['dado'])
     vazoes["aberturas"] = []
 
     # número de aberturas do último estágio
     for i in range(vazoes["numEstagios"]):
         if vazaoData[2+i]['dado'] != 1 and i != vazoes["numEstagios"]-1:
             raise RuntimeError("Número de aberturas diferente de 1 do estágio "+str(i+1))
-        vazoes["aberturas"].append(vazaoData[2+i]['dado'])
+        vazoes["aberturas"].append(int(vazaoData[2+i]['dado']))
         
 
     # registro 2 - codigo das usinas consideradas
     vazoes["codUhe"] = []
     for i in range(vazoes["numPostos"]):
-        vazoes["codUhe"].append(vazaoData[blockSize+i]['dado'])
+        vazoes["codUhe"].append(int(vazaoData[blockSize+i]['dado']))
 
 
     # registro 3 - número de semanas completas do estudo, nro de dias excluídos do estagio \
     # seguinte ao mes inicial,
 
-    vazoes["semanasCompletas"] = vazaoData[blockSize*2]['dado']
-    vazoes["numDiasExcl"] = vazaoData[blockSize*2+1]['dado']
-    vazoes["mesIni"] = vazaoData[blockSize*2+2]['dado']
-    vazoes["anoIni"] = vazaoData[blockSize*2+3]['dado']
+    vazoes["semanasCompletas"] = int(vazaoData[blockSize*2]['dado'])
+    vazoes["numDiasExcl"] = int(vazaoData[blockSize*2+1]['dado'])
+    vazoes["mesIni"] = int(vazaoData[blockSize*2+2]['dado'])
+    vazoes["anoIni"] = int(vazaoData[blockSize*2+3]['dado'])
 
     # registro 4 - probabilidades (pula)
 
@@ -43,7 +43,7 @@ def importVAZOES(fn,blockSize):
     for estagio in range(1,vazoes["semanasCompletas"]+1):
         vazoes["prevSem"].append([])
         for posto in range(1,blockSize+1):
-            vazoes["prevSem"][estagio-1].append(vazaoData[skip+pos]['dado'])
+            vazoes["prevSem"][estagio-1].append(int(vazaoData[skip+pos]['dado']))
             pos += 1
 
     return vazoes
